A couple of fixes.

+++ ext/standard/string.c
@@ -1889,7 +1889,7 @@ int php_tag_find(char *tag, int len, char *set) {
 			done =1;
 			break;
 		default:
-			if(!isspace(c)) {
+			if(!isspace((int)c)) {
 				if(state==0) {
 					state=1;
 					if(c!='/') *(n++) = c;
